tests/wpt/tests/ambient-light/AmbientLightSensor-supported-by-permissions-policy.html

<!DOCTYPE html>
<title>Test that ambient-light-sensor is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissionspolicy-features">
<link rel="help" href="https://w3c.github.io/sensors/#permissions-policy-api">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('ambient-light-sensor', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise ambient-light-sensor.');
</script>
